Split DatabaseEngine.cs into more files
